extract resolve method
